workers/requirements.txt
# Celery + broker
celery==5.4.0
kombu==5.3.7
flower==2.0.1
redis==5.0.7

# Database (sync — Celery workers are synchronous)
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
pgvector==0.3.2

# LLM
anthropic==0.34.0
openai==1.40.0

# Embeddings
sentence-transformers==3.0.1
torch==2.3.1

# Vector store
qdrant-client==1.10.1

# HTTP / scraping / auto-apply
httpx==0.27.0
requests==2.32.3
feedparser==6.0.11
beautifulsoup4==4.12.3
playwright==1.45.0

# Object storage
minio==7.2.7

# Config / utilities
pydantic==2.8.2
pydantic-settings==2.3.4
python-dotenv==1.0.1
pyyaml==6.0.2
tenacity==8.5.0
structlog==24.4.0
arrow==1.3.0
jinja2==3.1.4
